+++ src/main/java/net/emirikol/golemancy/entity/goal/GolemReturnHomeGoal.java
@@ -2,11 +2,7 @@ package net.emirikol.golemancy.entity.goal;
 
 import net.emirikol.golemancy.entity.*;
 
-import net.minecraft.entity.*;
-import net.minecraft.entity.mob.*;
-import net.minecraft.entity.ai.pathing.*;
 import net.minecraft.entity.ai.goal.*;
-import net.minecraft.world.*;
 import net.minecraft.util.math.*;
 
 import java.util.*;
@@ -35,7 +31,7 @@ public class GolemReturnHomeGoal extends Goal {
 	}
 	
 	public void start() {
-		this.entity.getNavigation().startMovingTo((double)((float)this.targetPos.getX()) + 0.5D, (double)(this.targetPos.getY() + 1), (double)((float)this.targetPos.getZ()) + 0.5D, this.speed);
+		this.entity.getNavigation().startMovingTo((double)((float)this.targetPos.getX()) + 0.5D, (this.targetPos.getY() + 1), (double)((float)this.targetPos.getZ()) + 0.5D, this.speed);
 		this.tryingTime = 0;
 		this.safeWaitingTime = this.entity.getRandom().nextInt(this.entity.getRandom().nextInt(1200) + 1200) + 1200;
 	}
@@ -44,7 +40,7 @@ public class GolemReturnHomeGoal extends Goal {
 		if (!this.targetPos.isWithinDistance(this.entity.getPos(), this.getDesiredSquaredDistanceToTarget())) {
 			++this.tryingTime;
 			if (this.shouldResetPath()) {
-				this.entity.getNavigation().startMovingTo((double)((float)this.targetPos.getX()) + 0.5D, (double)this.targetPos.getY(), (double)((float)this.targetPos.getZ()) + 0.5D, this.speed);
+				this.entity.getNavigation().startMovingTo((double)((float)this.targetPos.getX()) + 0.5D, this.targetPos.getY(), (double)((float)this.targetPos.getZ()) + 0.5D, this.speed);
 			}
 		} else {
 			--this.tryingTime;
